CODE_OF_CONDUCT.MD
Adding a code of conduct to your project
In this article
Adding a code of conduct using a template
Adding a code of conduct manually
Adopt a code of conduct to define community standards, signal a welcoming and inclusive project, and outline procedures for handling abuse.

A code of conduct defines standards for how to engage in a community. It signals an inclusive environment that respects all contributions. It also outlines procedures for addressing problems between members of your project's community. For more information on why a code of conduct defines standards and expectations for how to engage in a community, see the Open Source Guide.

Before adopting a code of conduct for your project:

Research different codes of conduct designed for open source projects. Choose one that reflects your community's standards.
Consider carefully whether you are willing and able to enforce it.
You can add a code of conduct to your project by using a template or manually creating a custom code of conduct. Your code of conduct will be available either way, but "Code of conduct" will only be marked as complete in your repository's community profile if you use a template. If you use a code of conduct written by another person or organization, be sure to follow any attribution guidelines from the source. For more information about community profiles, see "About community profiles for public repositories."

You can create a default code of conduct for your organization or personal account. For more information, see "Creating a default community health file."